Extract frame and task replace actor logic.
